Added ClickDrag element that allows you to click and drag it
All ClickDrag element does is allows you to define a function that happens when
it is clicked, and then it displays a value that you assign to it. Additionally,
it is styled with the `col-resize` cursor so that it's clear you can click and
drag left and right. All of the logic for how to handle a click down and change
the value happens above this element.